django/deploymentversion/researchsub_web/requirements.txt
crossrefapi==1.3.0
cryptography==2.2.2
Django==1.11.15
mysql-connector-python==8.0.12
mysqlclient==1.3.13
python-dateutil==2.4.2
urllib3==1.22
